refactor(spark): implement shared HoodieCatalystPlansUtils bodies in base and family classes
Move the method bodies that are byte-identical across all six Spark version
modules into the BaseHoodieCatalystPlanUtils trait in hudi-spark-common
(unapplyResolvedTable, projectOverSchema, isRepairTable,
getRepairTableChildren, the four index-plan matchers,
unapplyInsertIntoStatement, and createProjectForByNameQuery, replacing the
dead 'None' default that every version overrode identically).

Add family-level bases for bodies identical only within a Spark major
version: HoodieSpark3CatalystPlanUtils in hudi-spark3-common
(unapplyUpdateAction with the 2-field UpdateAction,
extractJsonFromSerializedOffset with the pre-4.1 package) and
HoodieSpark4CatalystPlanUtils in hudi-spark4-common (7-field MergeIntoTable,
5-field LogicalRelation scan, error-class based failAnalysisForMIT and
failTableNotFound).

Per-version objects keep only genuine Spark API deltas: pattern arities of
MergeIntoTable/ScanOperation/LogicalRelation/UpdateAction, the SerializedOffset
package move in Spark 4.1, pre-error-class failure messages on 3.3, the
_LEGACY_ERROR_TEMP_2309 error class on 3.4, and the Spark 3.4-only
default-columns workaround in unapplyInsertIntoStatement. Adapter-facing
object names are unchanged.
